backend/requirements.txt
selenium-wire
pandas
pydantic
bs4
xlrd
uvicorn
fastapi
python-dotenv
python-multipart
openpyxl